@@ -35,16 +35,10 @@ namespace LateStartStudio.AdventureGame.Game
         }
 
         /// <inheritdoc />
-        public override int Width
-        {
-            get { return this.sprite.Width; }
-        }
+        public override int Width => this.sprite.Width;
 
         /// <inheritdoc />
-        public override int Height
-        {
-            get { return this.sprite.Height; }
-        }
+        public override int Height => this.sprite.Height;
 
         /// <inheritdoc />
         public override bool Interact(int x, int y)
